Changed type on a few variables in entityevent. Before they where in, but since both the setter and getter where using float I changed this layer to use float as well. This also fixes compilation warnings.

+++ src/client/event/entityevent.h
@@ -157,8 +157,8 @@ namespace PT
     class EntityMoveEvent : public EntityEvent
     {
 		public:
-			int walkDirection;
-			int turnDirection;
+			float walkDirection;
+			float turnDirection;
 
     public:
       EntityMoveEvent() : EntityEvent("EntityMoveEvent", true) {}
@@ -171,7 +171,7 @@ namespace PT
 		class EntityMoveToEvent : public EntityEvent
 		{
 		public:
-			int speed;
+			float speed;
 			csVector3 origin;
 			csVector3 destination;
